ai챗봇모델 변경으로 인한 프론트,백엔드, ai 로직 변경

+++ backend/matefarm/src/main/java/com/avengers/matefarm/diagnosis/dto/VisionDiagnosisResponse.java
@@ -3,7 +3,7 @@ package com.avengers.matefarm.diagnosis.dto;
 import java.util.List;
 
 public record VisionDiagnosisResponse(
-        Integer cropId,
+        Long cropId,
         String modelKey,
         BestResponse best,
         List<BestResponse> topK,
